Making the order flow bit more "failure prone"
EXample: it will now respond back with pending which means its not done yet. If it takes "longer than" 30s it will now be cleaned up by the cleanup old task and that will rollback all other transactions.

This is horrible code but good enough to have played with the idea